Create Product Sales Data Analysis
Using this project, companies are able to keep track of their sales. As part of this we’ll be visualizing Microsoft’s sample dataset to show a profit by product and sales. The data set was connected to Power BI Desktop so that irrelevant data removed before visualization and exporting of the filtered data as dashboards.